requirements.txt
asgiref==3.9.1
black==25.1.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
cloudinary==1.44.1
crispy-bootstrap5==2025.6
Django==5.2.5
django-cloudinary-storage==0.3.0
django-crispy-forms==2.4
django-phonenumber-field==8.1.0
gunicorn==23.0.0
idna==3.10
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
phonenumbers==9.0.12
platformdirs==4.3.8
psycopg2-binary==2.9.10
python-decouple==3.8
python-dotenv==1.1.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
whitenoise==6.9.0
